tests: create librbd API-only integration test suite
The QA client-upgrade test suite requires a librbd test
that is dynamically linked to librbd.  Since the current
ceph_test_librbd includes tests against the non-public API,
it is statically linked against librbd and thus cannot be
used to test a client upgrade scenario.

Backport: hammer
(cherry picked from commit 6fe94c8cbb924c31c1007e2d0c76f28ce9dbca57)
