+++ src/test/librbd/test_ImageWatcher.cc
@@ -1,6 +1,7 @@
 // -*- mode:C; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
 // vim: ts=8 sw=2 smarttab
 #include "test/librbd/test_fixture.h"
+#include "test/librbd/test_support.h"
 #include "include/int_types.h"
 #include "include/stringify.h"
 #include "include/rados/librados.h"
