+++ src/test/librbd/test_main.cc
@@ -7,15 +7,19 @@
 #include "global/global_init.h"
 #include <vector>
 
+extern void register_test_librbd();
+#ifdef TEST_LIBRBD_INTERNALS
 extern void register_test_image_watcher();
 extern void register_test_internal();
-extern void register_test_librbd();
+#endif // TEST_LIBRBD_INTERNALS
 
 int main(int argc, char **argv)
 {
+  register_test_librbd();
+#ifdef TEST_LIBRBD_INTERNALS
   register_test_image_watcher();
   register_test_internal();
-  register_test_librbd();
+#endif // TEST_LIBRBD_INTERNALS
 
   ::testing::InitGoogleTest(&argc, argv);
